refactor isoformat to utils
